data-ipa-cfg-mgr:ipacm: fix the vlan id data type
vlan id can be more that 256. so increased the size
of vlan id to uint16_t.

+++ ipacm/inc/IPACM_ConntrackListener.h
@@ -69,7 +69,7 @@ typedef struct __nat_client_info
 	uint32_t nat_iface_ipv4_addr;
 #ifdef FEATURE_VLAN_MPDN
 	bool is_vlan_client;
-	uint8_t vlan_id;
+	uint16_t vlan_id;
 #endif
 }nat_client_info;
 
@@ -77,7 +77,7 @@ typedef struct __nat_client_info
 typedef struct _nat_pdn_entry
 {
 	uint32_t public_ip;
-	uint8_t vlan_id;
+	uint16_t vlan_id;
 }nat_pdn_entry;
 #endif
 
@@ -172,7 +172,7 @@ public:
 	void HandleSTAClientDelEvt_v6(const IpAddress& ip);
 	void ReadNfConntrackAcct();
 #ifdef FEATURE_VLAN_MPDN
-	bool IsVlanIPv4(uint32_t ipv4_address, uint8_t *VlanId);
+	bool IsVlanIPv4(uint32_t ipv4_address, uint16_t *VlanId);
 #endif
 };
